feat: セクションタイトルウィジェットを他画面でも使用できるように変更、それに伴う修正
