--- reimi/frontend/reimi_app/lib/presentation/features/setting/widgets/section_title.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SectionTitle extends StatelessWidget {
-  const SectionTitle({super.key, required this.title});
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const SizedBox(height: 16),
-        Text(
-          title,
-          style: Theme.of(context).textTheme.titleSmall!.copyWith(
-                fontSize: 13,
-                fontWeight: FontWeight.w600,
-                color: Colors.black87,
-              ),
-        ),
-        const SizedBox(height: 8),
-      ],
-    );
-  }
-}
